+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.7
+
+- Added
+  - Improved `TOMLDecodeError` error messages.
+    Line and column are included when applicable.
+
 ## 0.2.6
 
 - Added